Fully functioning client/server pair
Place both files in the same directory.  Open a command prompt and go to that directory.  Compile both files as so:
javac Receiver.java
javac Sender.java

Once the two files are compiled run the commands:
java Receiver to see what the command line needs to be.
java Sender to see what the command line needs to be.

Open a second command shell and go to the same directory where the files are.
Copy a binary file to that directory.
In one shell start the Receiver
In the other shell start the Sender
The input file passed in to the Sender should show up in the same directory using the output file name that was passed into the Receiver.

Lots of comments in the code, not real complicated, definitely room for enhancement.  Should be good for a basic foundation to leverage off of.